+++ R/rhat.R
@@ -1,6 +1,6 @@
 #' R-hat
 #'
-#' Calculates the uncorrected, untransformed, univariate
+#' By default calculates the uncorrected, untransformed, univariate
 #' split R-hat (potential scale reduction factor) values.
 #'
 #' @param x An MCMC object.
@@ -9,6 +9,7 @@
 #' "term", "parameter" or "all".
 #' @param as_df A flag indicating whether to return the values as a
 #' data frame versus a named list.
+#' @param split A flag specifying whether to split the chains in two.
 #' @param bound flag specifying whether to bind mcmcrs objects by their chains before calculating rhat.
 #' @return The rhat value(s).
 #' @references
@@ -27,24 +28,32 @@ rhat <- function(x, ...) {
 
 #' @describeIn rhat R-hat for an mcarray object
 #' @export
-rhat.mcarray <- function(x, by = "all", as_df = FALSE, ...)
-  rhat(as.mcmcarray(x), by = by, as_df = as_df)
+rhat.mcarray <- function(x, by = "all", as_df = FALSE, split = TRUE, ...) {
+  check_unused(...)
+  rhat(as.mcmcarray(x), by = by, as_df = as_df, split = TRUE)
+}
 
 #' @describeIn rhat R-hat for an mcmc object
 #' @export
-rhat.mcmc <- function(x, by = "all", as_df = FALSE, ...)
-  rhat(as.mcmcr(x), by = by, as_df = as_df)
+rhat.mcmc <- function(x, by = "all", as_df = FALSE, split = TRUE, ...) {
+  check_unused(...)
+  rhat(as.mcmcr(x), by = by, as_df = as_df, split = split)
+}
 
 #' @describeIn rhat R-hat for an mcmc.list object
 #' @export
-rhat.mcmc.list <- function(x, by = "all", as_df = FALSE, ...)
-  rhat(as.mcmcr(x), by = by, as_df = as_df)
+rhat.mcmc.list <- function(x, by = "all", as_df = FALSE, split = TRUE, ...) {
+  check_unused(...)
+  rhat(as.mcmcr(x), by = by, as_df = as_df, split = split)
+}
 
 #' @describeIn rhat R-hat for an mcmcarray object
 #' @export
-rhat.mcmcarray <- function(x, by = "all", as_df = FALSE, ...) {
+rhat.mcmcarray <- function(x, by = "all", as_df = FALSE, split = TRUE, ...) {
+  check_unused(...)
   check_vector(by, c("all", "parameter", "term"), length = 1)
   check_flag(as_df)
+  check_flag(split)
 
   if(niters(x) < 4) {
     if(!as_df) {
@@ -56,7 +65,7 @@ rhat.mcmcarray <- function(x, by = "all", as_df = FALSE, ...) {
     return(tibble(parameter = "parameter", rhat = NA_real_))
   }
 
-  x <- split_chains(x)
+  if(split) x <- split_chains(x)
   x <- apply(x, 3:ndims(x), FUN = .rhat)
 
   if(!as_df) {
@@ -72,29 +81,31 @@ rhat.mcmcarray <- function(x, by = "all", as_df = FALSE, ...) {
 
 #' @describeIn rhat R-hat for an mcmcr object
 #' @export
-rhat.mcmcr <- function(x, by = "all", as_df = FALSE, ...) {
+rhat.mcmcr <- function(x, by = "all", as_df = FALSE, split = TRUE, ...) {
+  check_unused(...)
   parameters <- parameters(x)
-  x <- lapply(x, rhat, by = by, as_df = as_df)
+  x <- lapply(x, rhat, by = by, as_df = as_df, split = split)
   if(!as_df) {
     if (by != "all") return(x)
     return(max(unlist(x)))
   }
-  x <- Map(x, parameters, f = function(x, p) {parameters(x[[1]]) <- p; x})
+  x <- Map(x, parameters, f = function(x, p) { parameters(x[[1]]) <- p; x})
   x <- do.call(rbind, x)
-  if(by == "all")
+  if (by == "all")
     return(tibble(all = "all", rhat = max(x$rhat)))
   x
 }
 
 #' @describeIn rhat R-hat for an mcmcrs object
 #' @export
-rhat.mcmcrs <- function(x, by = "all", as_df = FALSE, bound = FALSE, ...) {
+rhat.mcmcrs <- function(x, by = "all", as_df = FALSE, split = !bound, bound = FALSE, ...) {
   check_flag(bound)
   check_unused(...)
 
   if(bound) {
+    x <- lapply(x, collapse_chains)
     x <- Reduce(bind_chains, x)
-    return(rhat(x, by = by, as_df = as_df))
+    return(rhat(x, by = by, as_df = as_df, split = split))
   }
-  lapply(x, rhat, by = by, as_df = as_df)
+  lapply(x, rhat, by = by, as_df = as_df, split = split)
 }