+++ NEWS.md
@@ -1,5 +1,6 @@
-- add subset and parameters for mcmcrs object
-- add bound = FALSE argument to rhat.mcmcrs, esr.mcmcrs and converged.mcmcrs
+- add split = TRUE argument to rhat and converged
+- define subset and parameters for mcmcrs object
+- add bound = FALSE argument to rhat.mcmcrs and converged.mcmcrs
 
 # mcmcr 0.0.3